requirements.txt
PyQt6
pyqtgraph
watchdog
fpdf2
